Add pytest tests and CI
