+++ pyproject.toml
@@ -7,3 +7,6 @@ requires-python = ">=3.12"
 dependencies = [
     "streamlit>=1.49.1",
 ]
+
+[project.optional-dependencies]
+test = ["pytest>=7.4"]
